// src/app/api/model/getLoansLoanIdTransactions.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements. See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership. The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { GetLoansLoanIdPaymentType } from './getLoansLoanIdPaymentType';
import { GetLoansLoanIdLoanRepaymentScheduleInstallmentData } from './getLoansLoanIdLoanRepaymentScheduleInstallmentData';
import { GetLoansLoanIdLoanChargePaidByData } from './getLoansLoanIdLoanChargePaidByData';
import { GetLoansLoanIdLoanTransactionEnumData } from './getLoansLoanIdLoanTransactionEnumData';
import { GetLoansLoanIdCurrency } from './getLoansLoanIdCurrency';
import { GetLoansLoanIdLoanTransactionRelation } from './getLoansLoanIdLoanTransactionRelation';
import { GetLoansLoanIdCodeValueData } from './getLoansLoanIdCodeValueData';
import { GetLoansLoanIdPaymentDetailData } from './getLoansLoanIdPaymentDetailData';


/**
 * Set of GetLoansLoanIdTransactions
 */
export interface GetLoansLoanIdTransactions { 
    accountId?: number;
    accountNumber?: string;
    amount?: number;
    bankNumber?: number;
    checkNumber?: number;
    currency?: GetLoansLoanIdCurrency;
    date?: string;
    dateFormat?: string;
    externalId?: string;
    feeChargesPortion?: number;
    fixedEmiAmount?: number;
    id?: number;
    interestPortion?: number;
    /**
     * List of GetLoansLoanIdLoanChargePaidByData
     */
    loanChargePaidByList?: Array<GetLoansLoanIdLoanChargePaidByData>;
    /**
     * List of GetLoansLoanIdLoanRepaymentScheduleInstallmentData
     */
    loanRepaymentScheduleInstallments?: Array<GetLoansLoanIdLoanRepaymentScheduleInstallmentData>;
    locale?: string;
    manuallyReversed?: boolean;
    netDisbursalAmount?: number;
    numberOfRepayments?: number;
    officeId?: number;
    officeName?: string;
    outstandingLoanBalance?: number;
    overpaymentPortion?: number;
    paymentDetailData?: GetLoansLoanIdPaymentDetailData;
    paymentTypeId?: number;
    /**
     * List of GetLoansLoanIdPaymentType
     */
    paymentTypeOptions?: Array<GetLoansLoanIdPaymentType>;
    penaltyChargesPortion?: number;
    possibleNextRepaymentDate?: string;
    principalPortion?: number;
    receiptNumber?: number;
    reversalExternalId?: string;
    reversedOnDate?: string;
    routingCode?: number;
    submittedOnDate?: string;
    transactionAmount?: number;
    transactionDate?: string;
    /**
     * List of GetLoansLoanIdLoanTransactionRelationData
     */
    transactionRelations?: Set<GetLoansLoanIdLoanTransactionRelation>;
    transactionType?: string;
    type?: GetLoansLoanIdLoanTransactionEnumData;
    unrecognizedIncomePortion?: number;
    /**
     * List of GetLoansLoanIdCodeValueData
     */
    writeOffReasonOptions?: Array<GetLoansLoanIdCodeValueData>;
}